fix: ViewModelにBuildContextを渡さないような修正を行いました。
